+++ src/dsp_tools/commands/ingest_xmlupload/create_resources/upload_xml.py
@@ -17,9 +17,12 @@ from dsp_tools.commands.xmlupload.models.ingest import BulkIngestedAssetClient
 from dsp_tools.commands.xmlupload.models.upload_clients import UploadClients
 from dsp_tools.commands.xmlupload.models.upload_state import UploadState
 from dsp_tools.commands.xmlupload.prepare_xml_input.list_client import ListClientLive
-from dsp_tools.commands.xmlupload.prepare_xml_input.ontology_client import OntologyClientLive
-from dsp_tools.commands.xmlupload.prepare_xml_input.prepare_xml_input import prepare_upload_from_root
-from dsp_tools.commands.xmlupload.prepare_xml_input.read_validate_xml_file import validate_and_parse
+from dsp_tools.commands.xmlupload.prepare_xml_input.prepare_xml_input import get_intermediary_lookups
+from dsp_tools.commands.xmlupload.prepare_xml_input.prepare_xml_input import prepare_upload_from_root_ingest
+from dsp_tools.commands.xmlupload.prepare_xml_input.read_validate_xml_file import (
+    parse_and_validate_with_xsd_transform_special_tags,
+)
+from dsp_tools.commands.xmlupload.prepare_xml_input.read_validate_xml_file import preliminary_validation_of_root
 from dsp_tools.commands.xmlupload.project_client import ProjectClientLive
 from dsp_tools.commands.xmlupload.upload_config import UploadConfig
 from dsp_tools.commands.xmlupload.xmlupload import execute_upload
@@ -55,7 +58,6 @@ def ingest_xmlupload(
 
     auth = AuthenticationClientLive(server=creds.server, email=creds.user, password=creds.password)
     con = ConnectionLive(creds.server, auth)
-
     config = UploadConfig(
         media_previously_uploaded=True,
         interrupt_after=interrupt_after,
@@ -63,10 +65,14 @@ def ingest_xmlupload(
         server=creds.server,
         shortcode=shortcode,
     )
-
-    ontology_client = OntologyClientLive(con=con, shortcode=shortcode, default_ontology=default_ontology)
     clients = _get_live_clients(con, config, auth)
-    transformed_resources, stash = prepare_upload_from_root(root=root, ontology_client=ontology_client, clients=clients)
+
+    preliminary_validation_of_root(root, con, config)
+
+    intermediary_lookups = get_intermediary_lookups(root=root, clients=clients)
+    transformed_resources, stash = prepare_upload_from_root_ingest(
+        root=root, default_ontology=default_ontology, intermediary_lookups=intermediary_lookups
+    )
     state = UploadState(
         pending_resources=transformed_resources,
         pending_stash=stash,
@@ -90,7 +96,7 @@ def _parse_xml_and_replace_filepaths(xml_file: Path) -> tuple[str, etree._Elemen
     Raises:
         InputError: if replacing file paths with internal asset IDs failed
     """
-    root, shortcode, default_ontology = validate_and_parse(xml_file)
+    root, shortcode, default_ontology = parse_and_validate_with_xsd_transform_special_tags(xml_file)
 
     logger.info(f"Validated and parsed the XML. {shortcode=:} and {default_ontology=:}")