+++ src/dsp_tools/xmllib/internal_helpers.py
@@ -15,13 +15,28 @@ from dsp_tools.xmllib.models.config_options import NewlineReplacement
 from dsp_tools.xmllib.value_converters import replace_newlines_with_tags
 
 
-def _like_string(value: Any) -> bool:
+def is_nonempty_value_internal(value: Any) -> bool:
+    """
+    This is a duplicate of value_checkers.is_nonempty_value(), to avoid circular imports.
+
+    Check if a value is not None-like
+    or that its string representation contains at least one of the following characters:
+
+    - ``\\p{S}`` = symbols and special characters
+    - ``\\p{P}`` = punctuation
+    - ``\w`` = all Unicode letters, numbers, and _
+
+    Args:
+        value: value of any type
+
+    Returns:
+        True if the value is not None-like and contains at least one of the above-mentioned characters
+    """
     if pd.isna(value):
         return False
-    value = str(value).strip()
-    if len(value) == 0:
-        return False
-    return bool(regex.search(r"\S", value, flags=regex.UNICODE))
+    if regex.search(r"[\p{S}\p{P}\w]", str(value), flags=regex.UNICODE):
+        return True
+    return False
 
 
 def check_and_create_richtext_string(
@@ -61,7 +76,7 @@ def check_richtext_before_conversion(value: Any, prop_name: str, res_id: str) ->
         prop_name: Property name
         res_id: Resource ID
     """
-    if not _like_string(value):
+    if not is_nonempty_value_internal(value):
         msg = f"Resource '{res_id}' has a richtext value that is not a string: Value: {value} | Property: {prop_name}"
         warnings.warn(DspToolsUserWarning(msg))
 
@@ -94,30 +109,6 @@ def check_and_fix_collection_input(value: Any, prop_name: str, res_id: str) -> l
             raise InputError(msg)
         case _:
             return [value]
-
-
-def escape_reserved_xml_chars(richtext: str, known_tags: list[str]) -> str:
-    """
-    This function escapes characters that are reserved in an XML.
-    The angular brackets around the known tags will not be escaped.
-
-    Args:
-        richtext: Text to be escaped
-        known_tags: tags that should remain XML tags
-
-    Returns:
-        Escaped string
-    """
-    known_tags_regex = "|".join(known_tags)
-    lookahead = rf"(?!/?({known_tags_regex})/?>)"
-    illegal_lt = rf"<{lookahead}"
-    lookbehind = rf"(?<!</?({known_tags_regex})/?)"
-    illegal_gt = rf"{lookbehind}>"
-    illegal_amp = r"&(?![#a-zA-Z0-9]+;)"
-    richtext = regex.sub(illegal_lt, "&lt;", richtext or "")
-    richtext = regex.sub(illegal_gt, "&gt;", richtext)
-    richtext = regex.sub(illegal_amp, "&amp;", richtext)
-    return richtext
 
 
 def unescape_reserved_xml_chars(richtext: str) -> str: