+++ src/dsp_tools/xmllib/serialise/serialise_file_value.py
@@ -46,7 +46,7 @@ def _serialise_file_value(value: AbstractFileValue, authorship_id: str, tag_name
 
 def _serialise_metadata(metadata: Metadata, authorship_id: str) -> dict[str, str]:
     attribs = {
-        "license": metadata.license,
+        "license": str(metadata.license),
         "copyright-holder": metadata.copyright_holder,
         # the authorship that is contained in the metadata is the raw form, but we need its ID
         "authorship-id": authorship_id,