+++ test/unittests/commands/xmlupload/test_prepare_xml_input.py
@@ -2,10 +2,10 @@
 
 from lxml import etree
 
-from dsp_tools.commands.xmlupload.prepare_xml_input.prepare_xml_input import _extract_authorships_from_xml
+from dsp_tools.commands.xmlupload.prepare_xml_input.prepare_xml_input import _get_authorship_lookup
 
 
-def test_extract_authorships_from_xml_with_authors():
+def test_get_authorship_lookup():
     root_str = """
     <knora>
         <authorship id="authorship_1">
@@ -29,7 +29,7 @@ def test_extract_authorships_from_xml_with_authors():
         "authorship_2": ["Nora Ammann"],
         "authorship_3": ["Nora Ammann", "Johannes Nussbaum"],
     }
-    result = _extract_authorships_from_xml(root)
+    result = _get_authorship_lookup(root)
     assert set(result.keys()) == set(expected.keys())
     for k, v in result.items():
         assert set(v) == set(expected[k])
@@ -37,5 +37,5 @@ def test_extract_authorships_from_xml_with_authors():
 
 def test_extract_authorships_from_xml_no_authors():
     root = etree.fromstring("<knora></knora>")
-    result = _extract_authorships_from_xml(root)
+    result = _get_authorship_lookup(root)
     assert result == {}